+++ tests/testthat/test-plot_percentile_raster.R
@@ -1,6 +1,0 @@
-context("test-plot_percentile_raster.R")
-
-test_that("plot_percentile_raster works", {
-  x <- plot_percentile_raster(rstack1)
-  expect_true(is.null(x))
-})